refactor: 도메인 디렉토리를 모델과 레포지토리로 분리
